pyproject.toml
[project]
name = "tiktoktechjam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "easyocr>=1.7.2",
  "fastapi[standard]>=0.116.1",
  "ffmpeg-python>=0.2.0",
  "google-adk>=1.13.0",
  "google-genai>=1.32.0",
  "numpy<2",
  "opencv-python>=4.11.0.86",
  "python-dotenv>=1.1.1",
  "python-multipart>=0.0.20",
  "soundfile>=0.13.1",
  "uvicorn>=0.35.0",
]
